# source/asr-candidate/docs/api/openapi.json
{
  "openapi": "3.1.0",
...
  "info": {
    "title": "云蝠 Voice Runtime 管理与观测接口",
    "version": "1.13.0",
    "description": "云蝠 Voice Runtime 管理API：保留FreeSWITCH迁移对照，优先面向语音智能体媒体链路。1.10新增真实G711处理能力状态、显式隔离压测模式和完整内容/采样验收。未实现能力保持明确边界，不宣称完整FreeSWITCH或生产万路认证。"
  },
  "servers": [
    {
      "url": "http://127.0.0.1:9080",
      "description": "本机示例；实际由active.admin.listen决定。"
    }
  ],
  "security": [],
...
  "paths": {
...
    "/v1/docs": {
      "get": {
        "operationId": "getDocs",
        "summary": "读取文档清单",
        "description": "文档版本1.0.0，应用仍0.3.0。SHA-256和字节数对应原始文件，条目数不等于兼容认证通过数。",
        "tags": [
          "文档发布"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/Host"
          }
        ],
        "responses": {
          "200": {
            "description": "成功。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocsManifest"
                }
              }
            }
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "500": {
            "$ref": "#/components/responses/Internal"
          },
          "503": {
            "description": "管理重请求并发已达8个，尚未读取正文或执行业务；Retry-After为1秒，按原逻辑请求重试。状态、停止、排空及保护策略不占重请求槽。64个管理TCP连接全部占用时，新连接仍可能在系统队列等待，轻量入口也不能绕过连接上限。",
            "headers": {
              "Retry-After": {
                "schema": {
                  "type": "string",
                  "const": "1"
                },
                "description": "建议一秒后重试。"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "x-freeswitch": {
          "interface": "无同名FreeSWITCH HTTP入口",
          "relationship": "internal_contract_only",
          "difference": "这是RustSwitch自有HTTP契约，不兼容原版ESL/fs_cli的传输、命令与输出。",
          "status": "implemented",
          "reference_url": "https://github.com/signalwire/freeswitch/tree/ef32e205295e29f034f1453ad245ba5efb07b94a"
        }
      }
    },
    "/v1/docs/openapi.json": {
      "get": {
        "operationId": "getOpenAPI",
        "summary": "下载OpenAPI 3.1描述",
        "description": "本接口规范不代表FreeSWITCH原始API或ESL已实现。",
        "tags": [
          "文档发布"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/Host"
          }
        ],
        "responses": {
          "200": {
            "description": "成功。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpenAPIFile"
                }
              }
            },
            "headers": {
              "Content-Disposition": {
                "schema": {
                  "type": "string",
                  "description": "inline文件名，由mime.FormatMediaType编码。"
                }
              },
              "Content-Length": {
                "schema": {
                  "type": "integer",
                  "description": "非负累计数或即时数量。",
                  "minimum": 0
                }
              }
            }
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "500": {
            "$ref": "#/components/responses/Internal"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "503": {
            "description": "管理重请求并发已达8个，尚未读取正文或执行业务；Retry-After为1秒，按原逻辑请求重试。状态、停止、排空及保护策略不占重请求槽。64个管理TCP连接全部占用时，新连接仍可能在系统队列等待，轻量入口也不能绕过连接上限。",
            "headers": {
              "Retry-After": {
                "schema": {
                  "type": "string",
                  "const": "1"
                },
                "description": "建议一秒后重试。"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "x-freeswitch": {
          "interface": "无同名FreeSWITCH HTTP入口",
          "relationship": "internal_contract_only",
          "difference": "这是RustSwitch自有HTTP契约，不兼容原版ESL/fs_cli的传输、命令与输出。",
          "status": "implemented",
          "reference_url": "https://github.com/signalwire/freeswitch/tree/ef32e205295e29f034f1453ad245ba5efb07b94a"
        }
      }
    },
    "/v1/docs/comparison.json": {
      "get": {
        "operationId": "getComparison",
        "summary": "下载完整接口对照",
        "description": "结合status、relationship、differences和verification解释结果，禁止把静态条目当差分验收。",
        "tags": [
          "文档发布"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/Host"
          }
        ],
        "responses": {
          "200": {
            "description": "成功。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Comparison"
                }
              }
            },
            "headers": {
              "Content-Disposition": {
                "schema": {
                  "type": "string",
                  "description": "inline文件名，由mime.FormatMediaType编码。"
                }
              },
              "Content-Length": {
                "schema": {
                  "type": "integer",
                  "description": "非负累计数或即时数量。",
                  "minimum": 0
                }
              }
            }
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "500": {
            "$ref": "#/components/responses/Internal"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "503": {
            "description": "管理重请求并发已达8个，尚未读取正文或执行业务；Retry-After为1秒，按原逻辑请求重试。状态、停止、排空及保护策略不占重请求槽。64个管理TCP连接全部占用时，新连接仍可能在系统队列等待，轻量入口也不能绕过连接上限。",
            "headers": {
              "Retry-After": {
                "schema": {
                  "type": "string",
                  "const": "1"
                },
                "description": "建议一秒后重试。"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "x-freeswitch": {
          "interface": "无同名FreeSWITCH HTTP入口",
          "relationship": "internal_contract_only",
          "difference": "这是RustSwitch自有HTTP契约，不兼容原版ESL/fs_cli的传输、命令与输出。",
          "status": "implemented",
          "reference_url": "https://github.com/signalwire/freeswitch/tree/ef32e205295e29f034f1453ad245ba5efb07b94a"
        }
      }
    },
    "/v1/docs/file": {
      "get": {
        "operationId": "getDocument",
        "summary": "读取允许清单内原始文档",
        "description": "返回原始MD/JSON/CSV/头文件/文本，不执行Markdown或HTML，不读取任意磁盘路径。",
        "tags": [
          "文档发布"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/Host"
          },
          {
            "name": "path",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "从documents[].path原样取得。",
              "minLength": 1,
              "maxLength": 240
            },
            "description": "仅允许恰好一个path；缺失/重复/未知参数/解码错误/绝对路径/..穿越/反斜线400；合法但不在清单404。"
          }
        ],
        "responses": {
          "200": {
            "description": "原始正文，文本带UTF-8声明；具体MIME由允许文件类型决定。",
            "content": {
              "text/markdown": {
                "schema": {
                  "type": "string"
                }
              },
              "application/json": {
                "schema": {}
              },
              "text/csv": {
                "schema": {
                  "type": "string"
                }
              },
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "headers": {
              "Content-Disposition": {
                "schema": {
                  "type": "string",
                  "description": "inline; filename=安全basename。"
                }
              },
              "Content-Length": {
                "schema": {
                  "type": "integer",
                  "description": "非负累计数或即时数量。",
                  "minimum": 0
                }
              }
            }
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "400": {
            "$ref": "#/components/responses/BadPath"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/Internal"
          },
          "503": {
            "description": "管理重请求并发已达8个，尚未读取正文或执行业务；Retry-After为1秒，按原逻辑请求重试。状态、停止、排空及保护策略不占重请求槽。64个管理TCP连接全部占用时，新连接仍可能在系统队列等待，轻量入口也不能绕过连接上限。",
            "headers": {
              "Retry-After": {
                "schema": {
                  "type": "string",
                  "const": "1"
                },
                "description": "建议一秒后重试。"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "x-freeswitch": {
          "interface": "无同名FreeSWITCH HTTP入口",
          "relationship": "internal_contract_only",
          "difference": "这是RustSwitch自有HTTP契约，不兼容原版ESL/fs_cli的传输、命令与输出。",
          "status": "implemented",
          "reference_url": "https://github.com/signalwire/freeswitch/tree/ef32e205295e29f034f1453ad245ba5efb07b94a"
        }
      }
    },
    "/v1/docs/export.zip": {
      "get": {
        "operationId": "getDocsExport",
        "summary": "下载完整文档ZIP",
        "description": "包含manifest.json和documents列出的全部文件；不写配置，不执行文档。",
        "tags": [
          "文档发布"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/Host"
          }
        ],
        "responses": {
          "200": {
            "description": "成功。",
            "content": {
              "application/zip": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            },
            "headers": {
              "Content-Disposition": {
                "schema": {
                  "type": "string",
                  "description": ""
                },
                "example": "attachment; filename=\"rustswitch-interface-docs.zip\""
              },
              "Content-Length": {
                "schema": {
                  "type": "integer",
                  "description": "非负累计数或即时数量。",
                  "minimum": 0
                }
              }
            }
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "500": {
            "$ref": "#/components/responses/Internal"
          },
          "503": {
            "description": "管理重请求并发已达8个，尚未读取正文或执行业务；Retry-After为1秒，按原逻辑请求重试。状态、停止、排空及保护策略不占重请求槽。64个管理TCP连接全部占用时，新连接仍可能在系统队列等待，轻量入口也不能绕过连接上限。",
            "headers": {
              "Retry-After": {
                "schema": {
                  "type": "string",
                  "const": "1"
                },
                "description": "建议一秒后重试。"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "x-freeswitch": {
          "interface": "无同名FreeSWITCH HTTP入口",
          "relationship": "internal_contract_only",
          "difference": "这是RustSwitch自有HTTP契约，不兼容原版ESL/fs_cli的传输、命令与输出。",
          "status": "implemented",
          "reference_url": "https://github.com/signalwire/freeswitch/tree/ef32e205295e29f034f1453ad245ba5efb07b94a"
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "description": "诊断文本；不要依赖具体错误字串编写控制逻辑。"
          }
        },
        "description": "",
        "required": [
          "error"
        ],
        "additionalProperties": false
      },
...
      "DocumentItem": {
        "type": "object",
        "properties": {
          "path": {
            "type": "string",
            "description": "允许下载的相对文件路径。"
          },
          "title": {
            "type": "string",
            "description": "文档在人机界面中显示的标题。"
          },
          "category": {
            "type": "string",
            "description": "发布清单中的分类，不能据此推断兼容状态。"
          },
          "format": {
            "type": "string",
            "description": "md/json/csv/h/txt等文件格式。"
          },
          "bytes": {
            "type": "integer",
            "description": "原始文件字节数。",
            "minimum": 0
          },
          "sha256": {
            "type": "string",
            "description": "原始文件SHA-256。",
            "pattern": "^[0-9a-f]{64}$"
          }
        },
        "description": "",
        "required": [
          "path",
          "title",
          "category",
          "format",
          "bytes",
          "sha256"
        ],
        "additionalProperties": false
      },
      "DocsManifest": {
        "type": "object",
        "properties": {
          "version": {
            "type": "string",
            "description": "文档包版本；独立于应用版本。",
            "pattern": "^\\d+\\.\\d+\\.\\d+$",
            "examples": [
              "1.5.0"
            ]
          },
          "reference_version": {
            "type": "string",
            "description": "对照的固定 FreeSWITCH 版本。",
            "enum": [
              "1.11.3"
            ]
          },
          "reference_commit": {
            "type": "string",
            "description": "对照源代码的完整提交标识。",
            "enum": [
              "ef32e205295e29f034f1453ad245ba5efb07b94a"
            ]
          },
          "documents": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DocumentItem"
            },
            "description": "允许逐项下载的完整文件清单。"
          },
          "counts": {
            "type": "object",
            "properties": {
              "http_operations": {
                "type": "integer",
                "description": "显式业务/文档operation数量，不含静态资源和HEAD，本版20。",
                "minimum": 0
              },
              "comparison_entries": {
                "type": "integer",
                "description": "目录条目数，不是已兼容数量。",
                "minimum": 0
              }
            },
            "description": "本包的显式 HTTP 操作数与对照目录条目数。",
            "required": [
              "http_operations",
              "comparison_entries"
            ],
            "additionalProperties": false
          },
          "scope": {
            "type": "string",
            "description": "文档包描述的范围。"
          },
          "notice": {
            "type": "string",
            "description": "未实现或未验证能力等阅读注意事项。"
          }
        },
        "description": "",
        "required": [
          "version",
          "reference_version",
          "reference_commit",
          "documents",
          "counts",
          "scope",
          "notice"
        ],
        "additionalProperties": false
      },
      "ComparisonEntry": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "全目录唯一的稳定条目标识。"
          },
          "category": {
            "type": "string",
            "description": "对照类别或关键语义类别。"
          },
          "name": {
            "type": "string",
            "description": "原版入口、配置项或验收项的显示名。"
          },
          "module": {
            "type": "string",
            "description": "所属模块；没有模块对应关系时为说明文本。"
          },
          "fs_interface": {
            "type": "string",
            "description": "FreeSWITCH 接口名称或接口位置。"
          },
          "fs_syntax": {
            "type": "string",
            "description": "固定源码提取的原始语法，或无命令语法时的说明。"
          },
          "fs_semantics": {
            "type": "string",
            "description": "FreeSWITCH 一侧的行为范围。"
          },
          "rustswitch_interface": {
            "type": "string",
            "description": "当前 RustSwitch 对应入口或缺失说明。"
          },
          "rustswitch_semantics": {
            "type": "string",
            "description": "当前 RustSwitch 实际行为及其范围。"
          },
          "status": {
            "type": "string",
            "description": "本条入口或定义状态；implemented 不自动等于 FreeSWITCH 兼容。",
            "enum": [
              "implemented",
              "partial",
              "export_only",
              "not_implemented",
              "internal_only",
              "not_verified"
            ]
          },
          "relationship": {
            "type": "string",
            "description": "对应关系类型，区分直接对照、相似用途、仅导出和无兼容入口等情况。",
            "enum": [
              "direct_comparison",
              "analogous_not_wire_compatible",
              "export_artifact_only",
              "no_compatible_entrypoint",
              "internal_contract_only",
              "acceptance_requirement_only"
            ]
          },
          "verification": {
            "type": "string",
            "description": "核对依据和仍需执行的差分验收。"
          },
          "source_url": {
            "type": "string",
            "description": "来源链接；本地验收定义允许空串。"
          },
          "source_path": {
            "type": "string",
            "description": "固定源码树或项目文档内的来源路径。"
          },
          "document_path": {
            "type": "string",
            "description": "相对 docs 根的说明文档路径，可包含 Markdown 锚点。"
          },
          "differences": {
            "type": "array",
            "items": {
              "type": "string",
              "description": "具体差异。"
            },
            "description": "需要保留或补齐的具体语义差异列表。"
          },
          "source_line": {
            "type": "integer",
            "description": "源码或文档行号。",
            "minimum": 1
          }
        },
        "description": "implemented可仅表示自有HTTP接口实现，须结合relationship及差分证据判断；不能换算兼容完成率。",
        "required": [
          "id",
          "category",
          "name",
          "module",
          "fs_interface",
          "fs_syntax",
          "fs_semantics",
          "rustswitch_interface",
          "rustswitch_semantics",
          "status",
          "relationship",
          "verification",
          "source_url",
          "source_path",
          "document_path",
          "differences",
          "source_line"
        ],
        "additionalProperties": false
      },
      "Comparison": {
        "type": "object",
        "properties": {
          "version": {
            "type": "string",
            "description": "对照文档格式版本。"
          },
          "reference_version": {
            "type": "string",
            "description": "对照的固定 FreeSWITCH 版本。",
            "enum": [
              "1.11.3"
            ]
          },
          "reference_commit": {
            "type": "string",
            "description": "固定 FreeSWITCH 源码提交。",
            "enum": [
              "ef32e205295e29f034f1453ad245ba5efb07b94a"
            ]
          },
          "scope": {
            "type": "string",
            "description": "目录范围及不构成兼容认证的说明。"
          },
          "summary": {
            "type": "object",
            "properties": {
              "total": {
                "type": "integer",
                "description": "entries 中的总条数。",
                "minimum": 0
              },
              "by_category": {
                "type": "object",
                "additionalProperties": {
                  "type": "integer",
                  "description": "非负累计数或即时数量。",
                  "minimum": 0
                },
                "description": "类别名到条数的映射。"
              },
              "by_status": {
                "type": "object",
                "additionalProperties": {
                  "type": "integer",
                  "description": "非负累计数或即时数量。",
                  "minimum": 0
                },
                "description": "实现状态到条数的映射。"
              },
              "baseline_counts": {
                "type": "object",
                "additionalProperties": {
                  "type": "integer",
                  "description": "非负累计数或即时数量。",
                  "minimum": 0
                },
                "description": "固定基线各类静态发现数量；不等于运行中已加载数量。"
              }
            },
            "description": "目录数量统计，不能当作兼容通过率。",
            "required": [
              "total",
              "by_category",
              "by_status",
              "baseline_counts"
            ],
            "additionalProperties": false
          },
          "entries": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ComparisonEntry"
            },
            "description": "完整对照条目列表。"
          }
        },
        "description": "",
        "required": [
          "version",
          "reference_version",
          "reference_commit",
          "scope",
          "summary",
          "entries"
        ],
        "additionalProperties": false
      },
      "OpenAPIFile": {
        "type": "object",
        "properties": {
          "openapi": {
            "type": "string",
            "description": "本文件使用的 OpenAPI 规范版本。",
            "enum": [
              "3.1.0"
            ]
          },
          "info": {
            "type": "object",
            "properties": {
              "title": {
                "type": "string",
                "description": "接口描述的标题。"
              },
              "version": {
                "type": "string",
                "description": "文档版本，当前 1.0.0。"
              },
              "description": {
                "type": "string",
                "description": "接口描述及适用范围，可省略。"
              }
            },
            "description": "标题、文档版本及可选说明。",
            "required": [
              "title",
              "version"
            ],
            "additionalProperties": true
          },
          "paths": {
            "type": "object",
            "description": "HTTP 路径及操作定义；完整字段遵循 OpenAPI 3.1。"
          },
          "components": {
            "type": "object",
            "description": "复用 schema、响应、参数和安全机制定义。"
          },
          "jsonSchemaDialect": {
            "type": "string",
            "description": "schema 默认使用的 JSON Schema 方言。"
          },
          "servers": {
            "type": "array",
            "items": {
              "type": "object"
            },
            "description": "示例服务地址；实际监听地址由运行配置决定。"
          }
        },
        "description": "",
        "required": [
          "openapi",
          "info",
          "paths"
        ],
        "additionalProperties": true
      },
...
    },
    "responses": {
      "Forbidden": {
        "description": "Host不是localhost/回环字面量，或写请求Origin/CSRF校验失败。",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
...
      "Internal": {
        "description": "持久化、内嵌文件或目录处理错误；不可假定请求已生效。",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "BadPath": {
        "description": "查询/路径格式不合法；文档file入口比FS文件入口采用更严格的参数规则。",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "NotFound": {
        "description": "路径语法合法，但文件不在对应目录或文档允许清单。",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      }
    },
    "parameters": {
      "Host": {
        "name": "Host",
        "in": "header",
        "required": true,
        "schema": {
          "type": "string",
          "description": ""
        },
        "description": "客户端通常自动设置。主机部分须为精确localhost或回环IP字面量，自定义域名即便解析到回环也拒绝。"
      },
...
    },
    "securitySchemes": {
      "LocalCSRF": {
        "type": "apiKey",
        "in": "header",
        "name": "X-RustSwitch-CSRF",
        "description": "读取/v1/config.csrf_token；重启更换，仅防跨站写入，不是远程认证。"
      }
    }
  },
...
}